create status-bar.ts
